refactoring agent_session
